バフスキルについて更新。 スキルの制限時間、クールタイム、バフの種類（３種類）を設定し、PlayerUseSkillのBuffUse関数を呼ぶことで使用するようにした。

+++ Assets/Script/PlayerAttackParam.cs
@@ -6,18 +6,18 @@ public class PlayerAttackParam : MonoBehaviour
 {
     /// <summary>playerの基本攻撃力</summary>
     [SerializeField]
-    int _attackDamage = 500;//関数で変更する。
+    float _attackDamage = 500;//関数で変更する。
 
-    int _keepAttackDamage = 0;
+    float _keepAttackDamage = 0;
 
-    int _minAttackDamage = 400;
+    float _minAttackDamage = 400;
 
-    int _maxAttackDamage = 600;
+    float _maxAttackDamage = 600;
 
-    public int AttackDamage { get => _attackDamage; set => _attackDamage = value; }
-    public int KeepAttackDamage { get => _keepAttackDamage; set => _keepAttackDamage = value; }
-    public int MinAttackDamage { get => _minAttackDamage; set => _minAttackDamage = value; }
-    public int MaxAttackDamage { get => _maxAttackDamage; set => _maxAttackDamage = value; }
+    public float AttackDamage { get => _attackDamage; set => _attackDamage = value; }
+    public float KeepAttackDamage { get => _keepAttackDamage; set => _keepAttackDamage = value; }
+    public float MinAttackDamage { get => _minAttackDamage; set => _minAttackDamage = value; }
+    public float MaxAttackDamage { get => _maxAttackDamage; set => _maxAttackDamage = value; }
 
     private void Start()
     {
